fix: guard loadConfig against entity-change race + rename to PascalCase
- loadConfig now captures the entity identity at call time via a ref
  and discards the result if the entity changed while the GET was
  in-flight, preventing stale config from being applied to the wrong
  entity.
- Renames resource-tabs.tsx to ResourceTabs.tsx to match the PascalCase
  naming convention used by every other component module in the project.
  Updates all four import sites.
