+++ src/components/FunctionsPanel.tsx
@@ -20,7 +20,7 @@ import {
     renderResourceTabContent,
     isResourceTabId,
     type ResourceTabId,
-} from '@/components/resource-tabs';
+} from '@/components/ResourceTabs';
 import type { ComponentTopic, Operation, Fault } from '@/lib/types';
 
 /** Host app object returned from /functions/{id}/hosts */
