+++ src/components/ResourceTabs.tsx
